requirements.txt
dgl-cu111
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
ogb